// frontend/src/hooks/modal/useModal/types.ts
export type ModalPosition = {
  top: number | 'auto'
  bottom: number | 'auto'
  left: number | 'auto'
  right: number | 'auto'
}
// ...
export type BuildModalStyleOptions = {
  modalPosition: ModalPosition
  isWaitingForMeasurement?: boolean | null
  origin?: PopoverOrigin
  anchorDimensions: {
    width: number | null
    height: number | null
  }
  widthAnchor?: 'parent' | 'window' | 'anchor' | number
  fullWidth?: boolean
  maxHeight?: number | string
  maxWidth?: number | string
  modalElementParent?: HTMLElement | null
  isHorizontalPosition?: boolean
}